refactor: Move csv parser from emission to framework module.
This enable reuse of the generic csv parser.
